feat: 회원 닉네임으로 검색 API

+++ src/main/java/net/watchbox/domain/box/facade/my/MyBoxContentFacade.java
@@ -80,7 +80,7 @@ public class MyBoxContentFacade {
         // 5. 응답
         return ContentPageResponse.builder()
                 .contentItemList(contentItemList)
-                .totalCount((long) contentItemList.size())
+                .totalCount(contentItemList.size())
 //                .totalPages()
 //                .currentPage()
                 .build();
